test roles with service creation

+++ db/migrate/20200626191158_create_services.rb
@@ -2,7 +2,7 @@ class CreateServices < ActiveRecord::Migration[5.2]
   def change
     create_table :services do |t|
       t.string :name
-      t.numeric :price
+      t.integer :price
       t.string :description
       t.string :image_url
